+++ lib/smith/workflow/split_step_persistence/subclass_boundary.rb
@@ -46,6 +46,10 @@ module Smith
 
           { ttl: ttl }
         end
+
+        def dispatch_store!(...)
+          DispatchBoundary.instance_method(:dispatch_store!).bind_call(self, ...)
+        end
       end
     end
   end